fix(enclave): fix compilation warnings in enclave code

+++ src/enclave/outside/EnclaveInterface.cpp
@@ -110,7 +110,7 @@ int32_t EnclaveInterface::executeFunction(faabric::Message& msg)
     // Convert to vector of char pointers for easier serialisation
     std::vector<char*> cArgv;
     cArgv.resize(argv.size());
-    for (int i = 0; i < argv.size(); i++) {
+    for (size_t i = 0; i < argv.size(); i++) {
         cArgv.at(i) = const_cast<char*>(argv.at(i).c_str());
     }
